requirements.txt
aiogram==2.14.3
aiohttp==3.7.4
aioredis==1.3.1
amqp==5.0.6
APScheduler==3.6.3
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
Babel==2.9.1
bcrypt==3.2.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
bs4==0.0.1
CacheControl==0.12.6
cachetools==4.2.2
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
channels==3.0.4
channels-redis==3.3.0
chardet==3.0.4
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cobble==0.1.3
colorama==0.4.4
colorful==0.5.4
constantly==15.1.0
croniter==1.0.15
cryptography==3.4.7
daphne==3.0.2
defusedxml==0.7.1
Django==3.2.5
django-admin-rangefilter==0.10.0
django-admirarchy==1.2.1
django-allauth==0.45.0
django-appconf==1.0.5
django-celery-beat==2.2.1
django-constance==2.8.0
django-cors-headers==3.7.0
django-crispy-forms==1.12.0
django-crontab==0.7.1
django-dbbackup==3.3.0
django-debug-toolbar==3.2.1
django-filter==23.1
django-image-cropping==1.7
django-log-to-telegram==2020.1.23
django-mailbox==4.8.2
django-phonenumber-field==5.2.0
django-property-filter==1.1.2
django-push-notifications==3.0.0
django-redis==5.0.0
django-rest-auth==0.9.5
django-seed==0.3.1
django-simple-history==3.0.0
django-timezone-field==4.2.1
django-treenode==0.17.0
django-webpush==0.3.5
djangorestframework==3.12.4
djangorestframework-recursive==0.1.2
docx2pdf==0.1.7
docxtpl==0.11.5
easy-thumbnails==2.8.5
et-xmlfile==1.1.0
fabric==2.6.0
Faker==19.3.0
fcm-django==1.0.1
firebase-admin==5.0.1
firebird-base==1.3.0
firebird-driver==1.3.2
flower==1.0.0
google-api-core==1.31.1
google-api-python-client==2.15.0
google-auth==1.34.0
google-auth-httplib2==0.1.0
google-cloud-core==1.7.2
google-cloud-firestore==2.2.0
google-cloud-storage==1.41.1
google-crc32c==1.1.2
google-resumable-media==1.3.3
googleapis-common-protos==1.53.0
grpcio==1.39.0
gunicorn==20.0.4
hiredis==2.0.0
http-ece==1.1.0
httplib2==0.19.1
humanize==3.10.0
hyperlink==21.0.0
idna==3.2
importlib-metadata==4.6.1
incremental==21.3.0
invoke==1.6.0
Jinja2==3.0.1
kombu==5.1.0
lxml==4.6.3
mammoth==1.4.16
MarkupSafe==2.0.1
msgpack==1.0.2
multidict==5.1.0
ntlm-auth==1.5.0
numpy==1.23.2
oauthlib==3.1.1
openpyxl==3.0.9
packaging==21.0
pandas==1.4.3
paramiko==2.7.2
pathlib2==2.3.6
phonenumbers==8.12.28
Pillow==8.3.1
ply==3.11
prettyprinter==0.18.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
proto-plus==1.19.0
protobuf==3.17.3
psycopg2-binary==2.9.1
py-vapid==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodomex==3.10.1
Pygments==2.9.0
PyHamcrest==2.0.2
PyJWT==2.1.0
pykerberos==1.2.1
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pypandoc==1.6.3
pyparsing==2.4.7
pypsexec==0.2.0
pysmi==0.3.4
pysnmp==4.4.12
pyspnego==0.1.6
pyTelegramBotAPI==4.1.1
python-crontab==2.5.1
python-dateutil==2.8.2
python-docx==0.8.11
python-telegram-bot==13.7
python3-openid==3.2.0
pytz==2021.1
pywebpush==1.9.4
pywinrm==0.4.2
redis==3.5.3
requests==2.26.0
requests-kerberos==0.12.0
requests-ntlm==1.1.0
requests-oauthlib==1.3.0
rsa==4.7.2
service-identity==21.1.0
six==1.16.0
smbprotocol==1.5.1
soupsieve==2.2.1
sqlparse==0.4.1
toposort==1.10
tornado==6.1
tqdm==4.61.2
transliterate==1.10.2
Twisted==20.3.0
txaio==21.2.1
typing-extensions==3.10.0.0
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
websockets==9.1
xmltodict==0.12.0
yarl==1.6.3
zipp==3.5.0
zope.interface==5.4.0
